First two patterns!
The first pattern is `Unwrapable`. This pattern allows to unify every type which
implements an unwrap method, which can fail.

The second pattern is `Apply`. It allows to apply methods which do not take
ownership of object with ownership. This can for instance allow users to create
`HashMap` with a very nice syntax. This features three methods. The first method
ignore the values returned by the provided closure, the second one returns it
and the third ensures that this type is `Unwrapable` and tries to unwrap it.

Note for later: check the correct spelling for unwrapable.